% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeta.R
\name{getBeta}
\alias{getBeta}
\title{Estimate parameters of the distribution of a*Beta(1, b) .}
\usage{
getBeta(x)
}
\arguments{
\item{x}{- vector of real positive numbers.}
}
\value{
estimates of a and b. For a, it is the range of x. For b, two estimates are provided: the MLE, and the method of moment (matching mean).
}
\description{
Estimate parameters of the distribution of a*Beta(1, b) .
}
\examples{
getBeta(as.numeric(dat_udca$entry.dt))
}
\references{
Hong Zhang, Jie Pu, Shibing Deng, Satrajit Roychoudhury, Haitao Chu and Douglas Robinson. "Study Duration Prediction for Clinical Trials with Time-to-Event Endpoints Using Mixture Distributions Accounting for Heterogeneous Population", arXiv:2401.00540.
}
