% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeibull.R
\name{getWeilbull}
\alias{getWeilbull}
\title{Estimate parameters of the Weibull survival distributions for each subgroup.}
\usage{
getWeilbull(dat)
}
\arguments{
\item{dat}{- data frame of three columns: 1) time: follow up time; 2) status: indicator, 0=no event, 1=event; 3) group, integer 1,...,n, for each subgroup of patients.}
}
\value{
shape and scale parameters of the Weibull distribution for each subgroup.
}
\description{
Estimate parameters of the Weibull survival distributions for each subgroup.
}
\examples{
# dat_udca already has time, status and group columns defined,
getWeilbull(dat_udca)
}
\references{
Hong Zhang, Jie Pu, Shibing Deng, Satrajit Roychoudhury, Haitao Chu and Douglas Robinson. "Study Duration Prediction for Clinical Trials with Time-to-Event Endpoints Using Mixture Distributions Accounting for Heterogeneous Population", arXiv:2401.00540.
}
